split engagement settings to different parts
